+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
-Version 0.4 - 2012-Sep-07
+Version 0.4 - 2012-Oct-23
 ============================
 * Overall refactoring of API and therefore incompatible with previous versions.
-* Support pypy_.
+* Support pypy_ and python 3.3.
 * Added the new :mod:`pulsar.utils.httpurl` module for HTTP tools and HTTP clients.
 * Refactored :class:`pulsar.Deferred` to be more compatible with twisted. You
   can add separate callbacks for handling errors.
@@ -13,7 +13,7 @@ Version 0.4 - 2012-Sep-07
   a dictionary.
 * Socket and WSGI :class:`pulsar.Application` are built on top of the new
   :class:`pulsar.AsyncSocketServer` framework class.
-* **214 regression tests**, **79% coverage**.
+* **216 regression tests**, **79% coverage**.
 
 Version 0.3 - 2012-May-03
 ============================